requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
psycopg2-binary
python-dotenv==1.0.1
gunicorn==21.2.0
Werkzeug==3.0.1
flask-swagger-ui==4.11.1
PyYAML==6.0.1

